feat: add re-recommend and create situation buttons, implement all recommend button, and define recommendation types
