+++ src/domains/situation-create/components/button/create-situation-button.tsx
@@ -1,19 +1,16 @@
 import { CheckCircle2, Loader2 } from "lucide-react";
-import React, { useState } from "react";
 
 type CreateSituationButtonProps = {
   isFormFilled: boolean;
-  onClick: () => void;
+  isCreating: boolean;
 };
 export default function CreateSituationButton({
   isFormFilled,
-  onClick,
+  isCreating,
 }: CreateSituationButtonProps) {
-  const [isCreating, setIsCreating] = useState(false);
   return (
     <button
-      type="button"
-      onClick={onClick}
+      type="submit"
       disabled={!isFormFilled || isCreating}
       className={`w-full flex justify-center items-center py-4 px-6 text-lg font-extrabold rounded-lg shadow-lg transition duration-300 ease-in-out transform hover:scale-[1.02] focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-red-500
       ${
