+++ src/domains/situation-create/types/recommendation-types.ts
